_checkRange(n) function, and making Elastic functions call each other.
To avoid duplicate code, I've moved the `if not 0 <= n <= 1` check to a
separate function, and called it everywhere. It avoids repeating the
same error string everywhere, and it is a bit more readable. It may be a
tiny bit slower, due to the additional function call, but it is
negligible.

The Elastic easing functions were the most complex ones in this module.
Default argument values were used to remove a bunch of conditionals.
Also, two of those functions were rewritten to call the first one, just
like the Bounce functions already do.